renamed token.py and fixed imports
